+++ crates/epoch-snark/Cargo.toml
@@ -27,6 +27,7 @@ rand_xorshift = { version = "0.2" }
 bench-utils = { git = "https://github.com/scipr-lab/zexe", features = ["print-trace"] }
 hex = "0.3.2"
 bls-gadgets = { path = "../bls-gadgets", features = ["test-helpers"] }
+bls-crypto = { path = "../bls-crypto", features = ["test-helpers"] }
 
 [lib]
 crate-type = ["lib", "staticlib"]
